feat: Implement word limit functionality across various components
- Added word limit enforcement for the description and title fields in AdminCalendar and PackagesContent.
- Updated the Testimonials component to include a comment textarea.
- Enhanced BookingContent to format booking dates and display additional booking details.
- Introduced a new EventsFeed component to manage event listings and modals.
- Created a ShareModal for sharing event links on social media platforms.
- Integrated word limit functionality in the ReviewModal for name input.
