+++ src/components/Testimonials.tsx
@@ -177,6 +177,7 @@ const Testimonials = () => {
                   Enter review
                 </label>
                 <textarea
+                
                   id="comment"
                   name="comment"
                   placeholder="150 words"
